Prevent concurrent job execution with Hangfire attribute
Added the `[DisableConcurrentExecution(timeoutInSeconds: 5 * 60)]`
attribute to the `ExecuteAsync` methods in `SyncContentLog`,
`SyncContents`, and `SyncMovieGenres` classes to prevent
concurrent execution across distributed instances, with a
10-minute timeout.

Retained the use of `SemaphoreSlim` for additional concurrency
control within the application. Enhanced documentation in the
`SyncContents` class to clarify the purpose of the `ExecuteAsync`
method and its balanced approach to syncing content.